+++ SecretColors/__init__.py
@@ -6,5 +6,5 @@
 #
 #
 from SecretColors.models.palette import Palette
-from SecretColors.models.colormap import ColorMap
+from SecretColors.cmaps.parent import ColorMap
 from SecretColors.models.objects import ColorWheel