+++ SecretColors/data/cmaps/colorblind.py
@@ -1,0 +1,74 @@
+#  Copyright (c) SecretBiology  2020.
+#
+#  Library Name: SecretColors
+#  Author: Rohit Suratekar
+#  Website: https://github.com/secretBiology/SecretColors
+#
+#
+# Acknowledgments:
+# cud: Color Universal Design (https://jfly.uni-koeln.de/color/)
+#
+# tol: Paul Tol's Palettes (https://personal.sron.nl/~pault/#sec:qualitative)
+
+CB_DATA = {
+    "cud": {
+        '8': ['#000000', '#e69f00', '#56b4e9', '#009e73', '#f0e442',
+              '#0072b2', '#d55e00', '#cc79a7'],
+        "type": "qual"
+    },
+    "tol_bright": {
+        "7": ["#4477aa", "#66ccee", "#228833", "#ccbb44", "#ee6677", "#aa3377",
+              "#bbbbbb"],
+        "type": "qual"
+    },
+    "tol_contrast": {
+        "5": ["#ffffff", "#ddaa33", "#bb5566", "#004488", "#000000"],
+        "type": "qual"
+    },
+    "tol_vibrant": {
+        "7": ["#0077bb", "#33bbee", "#009988", "#ee7733", "#cc3311", "#ee3377",
+              "#bbbbbb"],
+        "type": "qual"
+    },
+    "tol_mutated": {
+        "9": ["#332288", "#88ccee", "#44aa99", "#117733", "#999933",
+              "#ddcc77", "#cc6677", "#882255", "#aa4499"],
+        "type": "qual"
+    },
+    "tol_pale": {
+        "6": ["#bbccee", "#cceeff", "#ccddaa", "#eeeebb", "#ffcccc",
+              "#dddddd"],
+        "type": "qual"
+    },
+    "tol_dark": {
+        "6": ["#222255", "#225555", "#225522", "#666633", "#663333",
+              "#555555"],
+        "type": "qual"
+    },
+    "tol_light": {
+        "9": ["#77aadd", "#99ddff", "#44bb99", "#bbcc33", "#aaaa00",
+              "#eedd88", "#ee8866", "#ffaabb", "#dddddd"],
+        "type": "qual"
+    },
+    "tol_sunset": {
+        "11": ["#364b9a", "#4a7bb7", "#6ea6cd", "#98cae1", "#c2e4ef",
+               "#eaeccc", "#feda8b", "#fdb366", "#f67e4b", "#dd3d2d",
+               "#a50026"],
+        "type": "div"
+    },
+    "tol_bu_rd": {
+        "9": ["#2166ac", "#4393c3", "#92c5de", "#d1e5f0", "#f7f7f7",
+              "#fddbc7", "#f4a582", "#d6604d", "#b2118sb"],
+        "type": "div"
+    },
+    "tol_pr_gn": {
+        "9": ["#762a83", "#9970ab", "#c2a5cf", "#e7d4e8", "#f7f7f7", "#d9f0d3",
+              "#acd39e", "#5aae61", "#1b7837"],
+        "type": "div"
+    },
+    "tol_yl_or_br": {
+        "9": ["#ffffe5", "#fff7bc", "#fee391", "#fec44f", "#fb9a29", "#ec7014",
+              "#cc4c02", "#993404", "#662506"],
+        "type": "seq"
+    }
+}