+++ SecretColors/models/objects.py
@@ -124,7 +124,49 @@ class ColorWheel:
     """
     ColorWheel class is more 'scientific' than using
     :class:`SecretColors.Palette`. This provides very useful
-    and easy color manipulation tools.
+    and easy color manipulation tools. This class essentially mimics the
+    typical color wheel. You can 'rotate' the wheel in different directions
+    and axis to get appropriate colors. Following code shows the simplest use
+
+    .. code-block:: python
+
+        cw = ColorWheel("#fa4d56")  # Initialize your 'base color'
+        print(cw.color)  # Prints #fa4d56
+        cw.rotate_hue(180)  # Rotates Hue by 180 degree
+        print(cw.color) # Prints #4dfaf1 (which is on the opposite side of  color wheel_)
+        cw.rotate_hue(-180)  # Rotate Hue by -180 degree
+        print(cw.color) # Back to #fa4d56
+        cw.rotate_lightness(10) # Rotates Lightness by 10
+        print(cw.color) # Prints #fa5b63
+        cw.rotate_saturation(-50) # Rotate saturation by -50
+        print(cw.color)  # Prints #ef676e
+
+    .. tip::
+        When your provide negative rotation values, essentially you are
+        rotating wheel anti-clockwise. Essentially, 0-360 degree represents
+        0-100% of value. When you go above 360 or below 0, it will
+        automatically wrap it around.
+
+    Sometimes, rotation might be little confusing while dealing with
+    lightness. Hence we have special methods which are
+    user-friendly
+
+    .. code-block:: python
+
+        cw = ColorWheel("#fa4d56")  # Initialize your 'base color'
+        cw.make_darker(30)  # Make current color darker by 30%
+        print(cw.color)  # Prints #df0612, darker shade of #fa4d56
+        cw.make_lighter(10)  # Make current color lighter by 30%
+        print(cw.color)  # Prints #f80915, lighter shade of #df0612
+
+    You can perform infinite amount of manipulations. Only thing you should
+    remember that, ColorWheel will return the 'current' color. So each time
+    you perform manipulation, color will change. However, at anytime if you
+    want to reset color to the original color (which you used to initialize
+    the ColorWheel) you can simply use :func:`~SecretColors.ColorWheel.reset`.
+
+    There are many useful methods which you can use to find colors with
+    specific color harmony (like monochromatic, complementary, etc).
 
     """
 
@@ -149,8 +191,6 @@ class ColorWheel:
     @property
     def color(self) -> str:
         """Returns current color (which has all the manipulations)
-
-        >>> cw = ColorWheel("#")
         """
         return hsl_to_hex(self.hue, self.saturation, self.lightness)
 
@@ -169,19 +209,49 @@ class ColorWheel:
         return value / 360
 
     def rotate_hue(self, angle: float):
+        """
+        Rotates Hue with given angle
+
+        :param angle: Angle of rotation
+        :type angle: float
+        """
         self.hue = self._rotate(self.hue, angle)
 
     def rotate_saturation(self, angle: float):
+        """
+        Rotates Saturation with given angle
+
+        :param angle: Angle of rotation
+        :type angle: float
+        """
         self.saturation = self._rotate(self.saturation, angle)
 
     def rotate_lightness(self, angle: float):
+        """
+        Rotates Lightness with given angle
+
+        :param angle: Angle of ratation
+        :type angle: float
+        """
         self.lightness = self._rotate(self.lightness, angle)
 
     def make_darker(self, percentage: float):
+        """
+        Makes color darker by reducing lightness
+
+        :param percentage: Percentage change
+        :type percentage: float
+        """
         av = self.lightness * percentage / 100
         self.lightness -= av
 
     def make_lighter(self, percentage: float):
+        """
+        Makes color lighter by increasing lightness
+
+        :param percentage: Percentage change
+        :type percentage: float
+        """
         av = (1 - self.lightness) * percentage / 100
         self.lightness += av
 
@@ -196,16 +266,51 @@ class ColorWheel:
             return color_list
 
     def complementary(self, is_reversed: bool = False) -> list:
-        # Make in HSV space
+        """
+        Generates two complementary colors. Out of which, one will be your
+        current color
+
+        :param is_reversed: If True, return list will be reversed
+        :type is_reversed: bool
+        :return: List of two complementary colors
+        """
         return self._make_color_list(180, is_revered=is_reversed)
 
     def triadic(self, is_reserved: bool = False) -> list:
+        """
+        Generate triadic colors. Out of which, one will be your current
+        color
+
+        :param is_reserved: If True, return list will be reversed
+        :type is_reserved: bool
+        :return: List of 3 triadic colors
+        """
         return self._make_color_list(120, 240, is_revered=is_reserved)
 
     def tetradic(self, is_reversed: bool = False) -> list:
+        """
+        Generates tetradic colors. Out of which, one will be your current
+        color.
+
+        :param is_reversed: If True, return list will be reversed
+        :type is_reversed: bool
+        :return: List of 4 tetradic colors
+        """
         return self._make_color_list(90, 180, 270, is_revered=is_reversed)
 
     def analogous(self, loc="first", is_reversed: bool = False) -> list:
+        """
+        Generate list of analogous colors. Out of which, one will be your
+        current color. Location of where current color should be in these 3
+        colors can be decided by :paramref:`loc` argument.
+
+        :param loc: Location of current color in the analogous colors.
+            Available options: [first, middle, last]. (default: first)
+        :type loc: str
+        :param is_reversed: If True, return list will be reversed.
+        :type is_reversed: bool
+        :return: List of 3 analogous colors
+        """
         factor = 1
         if loc.strip().lower() in ["last", "l"]:
             factor = -1
@@ -231,12 +336,18 @@ class ColorWheel:
             return colors
 
     def text_color(self) -> str:
+        """Simply returns black or white color based on the color contrast
+        of back/white with current color. This will be useful when you are
+        writing text on any colored background.
+
+        :return: White of Black color (based on contrast)
+        """
         return text_color(self.color)
 
 
 def run():
-    from SecretColors.utils import rgb_to_hex, hex_to_rgb
-    r, g, b = 0.3, 0.5, 0.7
-    hex_color = rgb_to_hex(r, g, b)  # hex_color: #4c80b2
-    k = hex_to_rgb(hex_color)
-    print(tuple(round(x, 2) for x in k))
+    cw = ColorWheel("#fa4d56")  # Initialize your 'base color'
+    cw.make_darker(30)  # Make current color darker by 30%
+    print(cw.color)  # Prints #df0612, darker shade of #fa4d56
+    cw.make_lighter(10)  # Make current color lighter by 30%
+    print(cw.color)  # Prints #f80915, lighter shade of #df0612